🚚 Rename & separate controllers, add documentation for aggregations
