+++ backend/src/controllers/status.controller.ts
@@ -1,0 +1,33 @@
+import { Controller, Get, Route, Response, Tags  } from 'tsoa';
+import { HealthcheckResponse } from '../models/result';
+
+/**
+ * @swagger
+ * tags:
+ *   name: Check
+ *   description: Vérification status du backend
+ */
+@Route('')
+export class StatusController extends Controller {
+
+  /**
+   * Health check endpoint
+   * @summary Requête utilise pour vérifier le bon fonctionnement du backend
+   */
+  @Response<HealthcheckResponse>('200', 'OK')
+  @Tags('Check')
+  @Get('/healthcheck')
+  public msg(): HealthcheckResponse {
+    return { msg: 'OK' };
+  }
+
+  /**
+   * Backend version endpoint
+   * @summary Obtenir la version du backend
+   */
+  @Tags('Check')
+  @Get('/version')
+  public version(): string {
+    return process.env.APP_VERSION;
+  }
+}
