+++ backend/src/server.ts
@@ -6,8 +6,10 @@ import loggerStream from './logger';
 // import { router as bulk } from './controllers/bulk';
 import { router as documentation } from './controllers/documentation';
 // Manually telling tsoa which controllers to use in the app entry file, route generation faster
-import "./controllers/index.controller";
+import "./controllers/search.controller";
 import "./controllers/bulk.controller";
+import "./controllers/aggregation.controller";
+import "./controllers/status.controller";
 
 export const app = express();
